+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.5.0] — 2019-06-27
+
 ### Added
 
 - [Oban.Pruner] Added `prune_limit` option to constrain the number of rows
@@ -167,7 +169,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - [Oban] Initial release with base functionality.
 
-[Unreleased]: https://github.com/sorentwo/oban/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/sorentwo/oban/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/sorentwo/oban/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/sorentwo/oban/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/sorentwo/oban/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/sorentwo/oban/compare/v0.1.0...v0.2.0